Adding valgrind testing to travis, and some changes to make it pass: (#334)
- `DataChunk` now always initialises its data
- Trigonometric functions do not execute on NULL payloads because of undefined behavior
- No more `NULL` values for empty `FROM` clauses, instead an `EmptyTableRef`/`BoundEmptyTableRef` is produced.

Issue #328

+++ src/common/types/chunk_collection.cpp
@@ -81,7 +81,8 @@ static int8_t templated_compare_value(Vector &left_vec, Vector &right_vec, index
 	return 1;
 }
 
-static int8_t compare_value(Vector &left_vec, Vector &right_vec, index_t vector_idx_left, index_t vector_idx_right) {
+// return type here is int32 because strcmp() on some platforms returns rather large values
+static int32_t compare_value(Vector &left_vec, Vector &right_vec, index_t vector_idx_left, index_t vector_idx_right) {
 
 	auto left_null = left_vec.nullmask[vector_idx_left];
 	auto right_null = right_vec.nullmask[vector_idx_right];
